Renamed helpers package to 'resource'
