+++ pkg/engine/resource/object_key_test.go
@@ -1,4 +1,4 @@
-package helpers
+package resource
 
 import (
 	"testing"
